Finish title + all/farm tabs header
